# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='700px' src='https://nitslaszlo.github.io/static/session/db_diagram_new.jpg' />"
    },
    "servers": [
        {
            "url": "https://jedliksession.cyclic.app"
        },
        {
            "url": "http://localhost:5000"
        }
    ],
    "paths": {
...
        "/authors": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Összes szerző (user, post páros) lekérdezése.",
                "description": "Összes szerző (user, post páros) lekérdezése.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/author"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "authors"
                ],
                "summary": "Új author mentése.",
                "description": "Új author mentése a 'post' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett author-al tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/author"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "name": []
        }
    ],
    "components": {
        "schemas": {
...
            "author": {
                "type": "object",
                "required": [
                    "user_id",
                    "post_id"
                ],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - Az author azonosítója (ObjectId).",
                        "example": "642747f1a0ef99fc1e2aeaf5"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "FK - A user azonosítója (ObjectId).",
                        "example": "61b5ea14f39e4edcf5b8a3bd"
                    },
                    "post_id": {
                        "type": "string",
                        "description": "FK - A pot azonosítója (ObjectID).",
                        "example": "61b6332df39e4edcf5b8a407"
                    }
                }
            },
...
            "ErrorSchema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "error.message value"
                    },
                    "status": {
                        "type": "integer",
                        "example": 400
                    }
                }
            }
        }
    }
}